0.2.0 release
- Somehow the checkout went wrong, now its a clone of the 0.0.10 from GrumpyOldPizza
- Removed unused Tlera boards
- Removed unused NUCLEO boards
- Altered B-L072Z-LRWAN1 board for use USB CDC for UART and programming
- Set VID/PID to 0x1209/0x0001 (pidcodes.io) for testing, should be altered for a right one

+++ cores/arduino/FS.cpp
@@ -118,7 +118,7 @@ int File::read() {
     return -1;
 }
 
-int File::read(uint8_t* data, size_t size) {
+size_t File::read(uint8_t* data, size_t size) {
     if (_handle) {
 	return _handle->read(data, size);
     }
